Created DTO to entities: Author, Availability and Book. Created mappers to map DTO to entities. Started creating services

+++ src/main/java/com/philippzeppelin/mdalib/database/entity/Book.java
@@ -19,11 +19,11 @@ public class Book {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String title;
 
-    private int publicationYear;
+    private Integer publicationYear;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "author_id")
